Rejoins settings after failure to be able to seperate them without big rewrites
