--- web/src/app/pages/settings/settings-credentials/settings-credentials.component.html
@@ -1,19 +1,0 @@
-<mat-expansion-panel>
-    <mat-expansion-panel-header>
-        <mat-panel-title>
-            Credentials
-        </mat-panel-title>
-        <mat-panel-description>
-            User Settings
-            <mat-icon>attach_money</mat-icon>
-        </mat-panel-description>
-    </mat-expansion-panel-header>
-    <form class="form-content">
-        <mat-form-field>
-            <input matInput name="username" [(ngModel)]="settings.SMSGlobal.Username" placeholder="Username">
-        </mat-form-field>
-        <mat-form-field>
-            <input matInput name="password" [(ngModel)]="settings.SMSGlobal.Password" placeholder="Password">
-        </mat-form-field>
-    </form>
-</mat-expansion-panel>
