--- web/src/app/pages/settings/settings-exchanges/settings-exchanges.component.html
@@ -1,29 +1,0 @@
-<mat-expansion-panel *ngFor="let exchange of settings?.Exchanges">
-    <mat-expansion-panel-header>
-        <mat-panel-title>
-            {{exchange.Name}}
-        </mat-panel-title>
-        <mat-panel-description>
-            Exchange Settings
-            <mat-icon>attach_money</mat-icon>
-        </mat-panel-description>
-    </mat-expansion-panel-header>
-    <form class="form-content">
-        <mat-checkbox name="exchangeEnabled" [(ngModel)]="exchange.Enabled">Enabled</mat-checkbox>
-        <div mat-line></div>
-        <mat-form-field>
-            <input matInput name="apiKey" [(ngModel)]="exchange.APIKey" [disabled]="!exchange.Enabled" placeholder="API Key*">
-        </mat-form-field>
-        <mat-form-field>
-            <input matInput name="apiSecretKey" [(ngModel)]="exchange.APISecret" [disabled]="!exchange.Enabled" placeholder="API Secret Key*">
-        </mat-form-field>
-        <mat-form-field>
-            <input matInput name="apiClientId" [(ngModel)]="exchange.ClientID" [disabled]="!exchange.Enabled" placeholder="API ClientID">
-        </mat-form-field>
-        <div mat-line></div>
-        <h4>
-            Enabled currencies
-        </h4>
-        <button mat-button [disabled]="!exchange.Enabled" (click)="openModal(exchange.Pairs)">CHANGE</button>
-    </form>
-</mat-expansion-panel>
