--- web/src/app/pages/settings/settings-exchanges/settings-exchanges.component.scss
@@ -1,12 +1,0 @@
-// FAB
-.mat-fab {
-    top: auto;
-    right: 30px;
-    bottom: 20px;
-    left: auto;
-    position: fixed;
-}
-
-.form-content {
-    margin: 20px;
-}
