--- web/src/app/pages/settings/settings-sms/settings-sms.component.html
@@ -1,32 +1,0 @@
-<mat-expansion-panel>
-    <mat-expansion-panel-header>
-        <mat-panel-title>
-            SMS Notifications
-        </mat-panel-title>
-        <mat-panel-description>
-            Notification Settings
-            <mat-icon>attach_money</mat-icon>
-        </mat-panel-description>
-    </mat-expansion-panel-header>
-    <form class="form-content">
-        <mat-checkbox name="smsEnabled" [(ngModel)]="settings.SMSGlobal.Enabled">Enabled</mat-checkbox>
-        <div mat-line></div>
-        <mat-form-field>
-            <input matInput name="smsUsername" [(ngModel)]="settings.SMSGlobal.Username" [disabled]="!settings?.SMSGlobal.Enabled" placeholder="Username">
-        </mat-form-field>
-        <mat-form-field>
-            <input matInput name="smsPassword" [(ngModel)]="settings.SMSGlobal.Password" [disabled]="!settings?.SMSGlobal.Enabled" placeholder="Password">
-        </mat-form-field>
-
-        <div *ngFor="let contact of settings.SMSGlobal?.Contacts">
-            <mat-checkbox name="contactEnabled" [disabled]="!settings?.SMSGlobal.Enabled" [(ngModel)]="contact.Enabled">Enabled</mat-checkbox>
-            <div mat-line></div>
-            <mat-form-field>
-                <input matInput name="contactUsername" [(ngModel)]="contact.Name" [disabled]="!settings.SMSGlobal.Enabled || !contact.Enabled" placeholder="Contact Name" />
-            </mat-form-field>
-            <mat-form-field>
-                <input matInput name="contactPassword" [(ngModel)]="contact.Number" [disabled]="!settings.SMSGlobal.Enabled || !contact.Enabled" placeholder="Contact Number" />
-            </mat-form-field>
-        </div>
-    </form>
-</mat-expansion-panel>
