--- web/src/app/pages/settings/settings-wallets/settings-wallets.component.html
@@ -1,12 +1,0 @@
-<mat-expansion-panel>
-    <mat-expansion-panel-header>
-        <mat-panel-title>
-            Wallets
-        </mat-panel-title>
-        <mat-panel-description>
-            Wallet Settings
-            <mat-icon>attach_money</mat-icon>
-        </mat-panel-description>
-    </mat-expansion-panel-header>
-
-</mat-expansion-panel>
