+++ web/src/app/pages/settings/settings.component.html
@@ -1,12 +1,107 @@
 <div class="loading-spinner" *ngIf="settings === null">
     <mat-progress-spinner color="accent" mode="indeterminate"></mat-progress-spinner>
 </div>
-<div *ngIf="settings !== null">
+<div *ngIf="settings !== null && ready">
     <button matTooltip="Save" (click)="saveSettings()" mat-fab color="accent" class="mat-fab mat-fab-bottom-right"><mat-icon >save</mat-icon></button>
     <mat-accordion>
-        <app-settings-credentials ([settings])="settings"></app-settings-credentials>
-        <app-settings-sms ([settings])="settings"></app-settings-sms>
-        <app-settings-wallets ([settings])="settings"></app-settings-wallets>
-        <app-settings-exchanges ([settings])="settings"></app-settings-exchanges>
+        <mat-expansion-panel>
+            <mat-expansion-panel-header>
+                <mat-panel-title>
+                    Credentials
+                </mat-panel-title>
+                <mat-panel-description>
+                    User Settings
+                    <mat-icon>person</mat-icon>
+                </mat-panel-description>
+            </mat-expansion-panel-header>
+            <form class="form-content">
+                <mat-form-field>
+                    <input matInput name="username" placeholder="Username">
+                </mat-form-field>
+                <mat-form-field>
+                    <input matInput name="password"  placeholder="Password">
+                </mat-form-field>
+            </form>
+        </mat-expansion-panel>
+        
+        <mat-expansion-panel>
+            <mat-expansion-panel-header>
+                <mat-panel-title>
+                    SMS Notifications
+                </mat-panel-title>
+                <mat-panel-description>
+                    Notification Settings
+                    <mat-icon>sms</mat-icon>
+                </mat-panel-description>
+            </mat-expansion-panel-header>
+            <form class="form-content">
+                <mat-checkbox name="smsEnabled" [(ngModel)]="settings.SMSGlobal.Enabled">Enabled</mat-checkbox>
+                <div mat-line></div>
+                <mat-form-field>
+                    <input matInput name="smsUsername" [(ngModel)]="settings.SMSGlobal.Username" [disabled]="!settings?.SMSGlobal.Enabled" placeholder="Username">
+                </mat-form-field>
+                <mat-form-field>
+                    <input matInput name="smsPassword" [(ngModel)]="settings.SMSGlobal.Password" [disabled]="!settings?.SMSGlobal.Enabled" placeholder="Password">
+                </mat-form-field>
+        
+                <div *ngFor="let contact of settings.SMSGlobal?.Contacts">
+                    <mat-checkbox name="contactEnabled" [disabled]="!settings?.SMSGlobal.Enabled" [(ngModel)]="contact.Enabled">Enabled</mat-checkbox>
+                    <div mat-line></div>
+                    <mat-form-field>
+                        <input matInput name="contactUsername" [(ngModel)]="contact.Name" [disabled]="!settings.SMSGlobal.Enabled || !contact.Enabled" placeholder="Contact Name" />
+                    </mat-form-field>
+                    <mat-form-field>
+                        <input matInput name="contactPassword" [(ngModel)]="contact.Number" [disabled]="!settings.SMSGlobal.Enabled || !contact.Enabled" placeholder="Contact Number" />
+                    </mat-form-field>
+                </div>
+            </form>
+        </mat-expansion-panel>
+
+
+        <mat-expansion-panel>
+    <mat-expansion-panel-header>
+        <mat-panel-title>
+            Wallets
+        </mat-panel-title>
+        <mat-panel-description>
+            Wallet Address Settings
+            <mat-icon>account_balance_wallet</mat-icon>
+        </mat-panel-description>
+    </mat-expansion-panel-header>
+    <mat-action-row>
+        <button mat-button>ADD</button>
+    </mat-action-row>
+</mat-expansion-panel>
+        
+        <mat-expansion-panel *ngFor="let exchange of settings?.Exchanges">
+            <mat-expansion-panel-header>
+                <mat-panel-title>
+                    {{exchange.Name}}
+                </mat-panel-title>
+                <mat-panel-description>
+                    Exchange Settings
+                    <mat-icon>poll</mat-icon>
+                </mat-panel-description>
+            </mat-expansion-panel-header>
+            <form class="form-content">
+                <mat-checkbox name="exchangeEnabled" [(ngModel)]="exchange.Enabled">Enabled</mat-checkbox>
+                <div mat-line></div>
+                <mat-form-field>
+                    <input matInput name="apiKey" [(ngModel)]="exchange.APIKey" [disabled]="!exchange.Enabled" placeholder="API Key*">
+                </mat-form-field>
+                <mat-form-field>
+                    <input matInput name="apiSecretKey" [(ngModel)]="exchange.APISecret" [disabled]="!exchange.Enabled" placeholder="API Secret Key*">
+                </mat-form-field>
+                <mat-form-field>
+                    <input matInput name="apiClientId" [(ngModel)]="exchange.ClientID" [disabled]="!exchange.Enabled" placeholder="API ClientID">
+                </mat-form-field>
+                <div mat-line></div>
+                <h4>
+                    Enabled currencies
+                </h4>
+                <button mat-button [disabled]="!exchange.Enabled" (click)="openModal(exchange.Pairs)">CHANGE</button>
+            </form>
+        </mat-expansion-panel>
+
     </mat-accordion>
 </div>
