+++ web/src/app/pages/settings/settings.component.scss
@@ -19,7 +19,3 @@
     height: 100%;
     width: 100%;
 }
-
-.form-content {
-    margin: 20px;
-}
